requirements.txt
# requirements.txt
# pip install -r requirements.txt


# Google Search Console support
google-api-python-client==2.176.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0

# GA4 SDK (compatible with Python 3.11+)
google-analytics-data==0.18.19

# Data handling
pandas==2.2.2

# CLI and utility tools
python-dateutil>=2.9.0
tqdm>=4.66.4

python-dotenv>=1.0.1